requirements.txt
# requirements.txt
#matplotlib==3.7.2
#openai==1.3.3
#python-dotenv==1.0.0

psycopg2-binary==2.9.9
openai==1.14.3
matplotlib==3.8.4
python-dotenv==1.0.1